refactor: move store/identityStore.ts to types folder
