+++ src/components/VerificationStatusBadge.tsx
@@ -1,5 +1,5 @@
 
-import { verifyStatuses } from "~/store/IdentityStore";
+import { verifyStatuses } from "~/types/Identity";
 import { Badge } from "./ui/badge";
 
 export function VerificationStatusBadge({ status }: { status: number }) {
