+++ src/hooks/useChallengeWebSocket.tsx
@@ -2,7 +2,7 @@ import { SS58String } from 'polkadot-api';
 import { useEffect, useCallback, useState, useRef } from 'react';
 import { AlertPropsOptionalKey } from '~/store/AlertStore';
 import { ChallengeStatus, ChallengeStore } from '~/store/challengesStore';
-import { IdentityInfo, verifyStatuses } from '~/store/IdentityStore';
+import { IdentityInfo, verifyStatuses } from '~/types/Identity';
 
 // Types matching your Rust backend
 type Data = {
